admin_panel/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["MATwave <stanislav.matveew@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
python-dotenv = "^1.0.0"
django-split-settings = "^1.2.0"
django-debug-toolbar = "^4.0.0"
django-extensions = "^3.2.1"
django-cors-headers = "^3.14.0"
psycopg2-binary = "^2.9.6"
dacite = "^1.8.0"
loguru = "^0.7.0"
pre-commit = "^3.3.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
bandit = "^1.7.5"
pyright = "^1.1.306"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
